Update: Copying libraries to disk on switching views
This is a temporary fix to allow the code view to show the actually libraries copied to disk, otherwise there is no current feedback to users to know where these files exists (or would) on disk until the operation is made. This allows the user to edit a file consuming libraries in memory to edit in code view and changes reflect such changes without losing track of libraries.